feat: support compile windows platform from source.

+++ bridge/core/page.cc
@@ -127,7 +127,7 @@ void WebFPage::evaluateScript(const char* script, size_t length, const char* url
 uint8_t* WebFPage::dumpByteCode(const char* script, size_t length, const char* url, uint64_t* byteLength) {
   if (!context_->IsContextValid())
     return nullptr;
-  return context_->DumpByteCode(script, length, url, byteLength);
+  return context_->DumpByteCode(script, static_cast<uint32_t>(length), url, byteLength);
 }
 
 bool WebFPage::evaluateByteCode(uint8_t* bytes, size_t byteLength) {
